+++ Adventure_Time/TextureManager.cpp
@@ -45,8 +45,8 @@ void TextureManager::load(const std::string& fileName, const std::string& id, SD
 }
 
 void TextureManager::drawSpritePic(const std::string& id, int x, int y,
-	int w, int h, SDL_Renderer* pRenderer, int indexFrame, float scope,
-	SDL_RendererFlip f, float a, SDL_Point* c)
+	int w, int h, SDL_Renderer* pRenderer, int indexFrame, double scope,
+	SDL_RendererFlip f, double a, SDL_Point* c)
 {
 	SDL_Rect srcRect = { w * indexFrame, 0, w, h };
 	w = static_cast<int>(w * scope);
